Check JSON response from node/read.
Previously was just using the XML return. Note that the test was
extended to also check that the keys of the top level JSON object
are in the order that we want them in.

Also added empty arrays for 'nodes', 'ways' and 'relations' to
match the format spec from https://github.com/osmlab/openstreetmap-json-schema/blob/gh-pages/1.0-DRAFT/Data.md